+++ src/frontend/apps/impress/src/features/docs/doc-management/api/useMoveDoc.tsx
@@ -2,7 +2,9 @@ import { TreeViewMoveModeEnum } from '@gouvfr-lasuite/ui-components';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { APIError, errorCauses, fetchAPI } from '@/api';
+import { KEY_DOC_TREE } from '@/docs/doc-tree/api/useDocTree';
 
+import { KEY_DOC } from './useDoc';
 import { KEY_LIST_DOC } from './useDocs';
 
 export type MoveDocParam = {
@@ -37,9 +39,9 @@ export function useMoveDoc() {
   return useMutation<void, APIError, MoveDocParam>({
     mutationFn: moveDoc,
     onSuccess() {
-      void queryClient.invalidateQueries({
-        queryKey: [KEY_LIST_DOC],
-      });
+      void queryClient.invalidateQueries({ queryKey: [KEY_LIST_DOC] });
+      void queryClient.invalidateQueries({ queryKey: [KEY_DOC] });
+      void queryClient.invalidateQueries({ queryKey: [KEY_DOC_TREE] });
     },
   });
 }